Organized the files by creating new directories and putting the appropriate files in them
