+++ app/src/main/java/com/example/spotifyclone/MainActivity.kt
@@ -19,6 +19,7 @@ package com.example.spotifyclone
 // The vast majority of source code I got was from the kotlin bootcamp with some help using chat gpt.
 // I mainly used chat gpt to aid with printing an actual bookItem when the add button is pressed, I
 // overshot with the state and user input criteria and I struggled to figure it out on my own
+import android.app.LauncherActivity.ListItem
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -92,8 +93,12 @@ import com.codelab.basics.R
 import com.example.spotifyclone.data.About
 import com.example.spotifyclone.data.questions
 import com.example.spotifyclone.ui.theme.SpotifyCloneTheme
+import com.example.spotifyclone.ui.books.BookCollection
+import com.example.spotifyclone.ui.settings.Settings
+import com.example.spotifyclone.ui.about.AboutUs
 import com.example.woof.data.Book
 import com.example.woof.data.books
+import java.util.Objects
 
 enum class Theme {
     Light, Dark
@@ -134,28 +139,16 @@ fun MyApp(modifier: Modifier = Modifier) {
 
                 ) {
                 Row {
-                    NavigationBarItem(selected = false, onClick = { navController.navigateSingleTopTo("settings")}, icon = {
-                        Icon(
-                            Icons.Rounded.Settings,
-                            contentDescription = "Home Icon",
-                            tint = Color.White
-                        )
-                    })
-                    NavigationBarItem(selected = false, onClick = {  navController.navigateSingleTopTo("bookCollection")}, icon = {
-                        Icon(
-                            Icons.Rounded.Book,
-                            contentDescription = "Book icon",
-                            tint = Color.White
-                        )
-                    })
-                    NavigationBarItem(selected = false, onClick = {  navController.navigateSingleTopTo("about")}, icon = {
-                        Icon(
-                            Icons.Rounded.Person,
-                            contentDescription = "About us icon",
-                            tint = Color.White
+                    ReadifyScreens.forEach { readifyDestination ->
 
-                        )
-                    })
+                        NavigationBarItem(selected = false, onClick = { navController.navigateSingleTopTo(readifyDestination.route) }, icon = {
+                            Icon(
+                                readifyDestination.icon,
+                                contentDescription = "${readifyDestination.route} icon",
+                                tint = Color.White
+                            )
+                        })
+                    }
                 }
             }
         }
@@ -181,7 +174,6 @@ fun MyApp(modifier: Modifier = Modifier) {
     }
 
 
-
 fun NavHostController.navigateSingleTopTo(route: String) =
     this.navigate(route) {
         popUpTo(
@@ -194,299 +186,16 @@ fun NavHostController.navigateSingleTopTo(route: String) =
     }
 
 
-@Composable
-fun BookCollection(modifier: Modifier = Modifier) {
-
-    val (bookList, setBookList) = remember { mutableStateOf(books) }
-    LazyColumn(modifier = modifier) {
-        item {
-            // Displays the "Add Book" card
-            AddBookItem(
-                onAddBookClick = { newBook ->
-                    setBookList(bookList + newBook)
-                }
-            )
-        }
-        items(bookList) { book ->
-            // Displays individual book items
-            BookItem(
-                book = book,
-                modifier = Modifier.padding(dimensionResource(R.dimen.padding_small))
-            )
-        }
-    }
-}
-
-@Composable
-fun Settings(modifier: Modifier = Modifier) {
-    var sliderValue by remember { mutableStateOf(0f) }
-    Surface(
-        modifier = modifier.fillMaxSize(),
-    ) {
-        Column() {
-            selectTheme()
-            Slider(
-                value = sliderValue,
-                onValueChange = {
-                    sliderValue = it
-                },
-                valueRange = 0f..100f, // Define the range for the slider
-                steps = 100 // Number of steps within the range
-            )
-        }
-
-    }
-}
-
-@Composable
-fun selectTheme(modifier: Modifier = Modifier) {
-    val radioOptions = listOf(Theme.Light, Theme.Dark)
-    var (selectedTheme, onThemeChange) = remember { mutableStateOf(Theme.Light ) }
-    Column {
-        radioOptions.forEach { theme ->
-            Row(
-                Modifier
-                    .fillMaxWidth()
-                    .selectable(
-                        selected = (theme == selectedTheme),
-                        onClick = {
-                            onThemeChange(theme)
-                        }
-                    )
-                    .padding(horizontal = 16.dp)
-            ) {
-                RadioButton(
-                    selected = (theme == selectedTheme),
-                    onClick = { onThemeChange(theme) }
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(text = theme.name, modifier = Modifier.padding(start = 16.dp))
-
-            }
-        }
-    }
-}
 
 
-/**
- * Composable function for adding a new book item.
- * @param onAddBookClick Callback when the "Add" button is clicked.
- */
-@Composable
-fun AddBookItem(
-    onAddBookClick: (Book) -> Unit,
-) {
-    // State variables for managing user input
-    var isExpanded by remember { mutableStateOf(false) }
-    var bookName by remember { mutableStateOf("") }
-    var releaseDate by remember { mutableStateOf("") }
-    var description by remember { mutableStateOf("") }
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { isExpanded = !isExpanded }
-            .padding(dimensionResource(R.dimen.padding_small))
-            .animateContentSize()
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                // Display the "Add Book" text
-                Text(
-                    text = "Add Book",
-                    modifier = Modifier.padding(dimensionResource(R.dimen.padding_small))
-                )
-            }
-
-            if (isExpanded) {
-                // User input fields
-                TextField(
-                    value = bookName,
-                    onValueChange = { bookName = it },
-                    label = { Text("Book Name") },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(dimensionResource(R.dimen.padding_small))
-                )
-
-                TextField(
-                    value = releaseDate,
-                    onValueChange = { releaseDate = it },
-                    label = { Text("Release Date") },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(dimensionResource(R.dimen.padding_small))
-                )
-
-                TextField(
-                    value = description,
-                    onValueChange = { description = it },
-                    label = { Text("Description") },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(dimensionResource(R.dimen.padding_small))
-                )
-
-                // Add button
-                Button(
-                    onClick = {
-                        // Validates user input and create a new Book object
-                        val newBook = createBook(bookName, releaseDate, description)
-                        if (newBook != null) {
-                            // Add the new book to the list
-                            onAddBookClick(newBook)
-
-                            // Clears input fields
-                            bookName = ""
-                            releaseDate = ""
-                            description = ""
-                        }
-                    },
-                    modifier = Modifier
-                        .align(Alignment.CenterHorizontally)
-                        .padding(dimensionResource(R.dimen.padding_small))
-                ) {
-                    Text("Add")
-                }
-            }
-        }
-    }
-}
-
-/**
- * Create a new Book object based on user input.
- * @param bookName The name of the book.
- * @param releaseDateStr The release date of the book as a string.
- * @param description The description of the book.
- * @return The created Book object or null if input is invalid.
- */
-private fun createBook(bookName: String, releaseDateStr: String, description: String): Book? {
-    val releaseDate = releaseDateStr.toIntOrNull()
-    if (releaseDate == null || bookName.isEmpty() || description.isEmpty()) {
-        // Handles validation errors here
-        return null
-    }
-
-    // Creates and return a new Book object
-    return Book(R.drawable.ph, bookName, releaseDate, description)
-}
-
-/**
- * Composable that displays a list item containing a book icon and its information.
- * @param book Contains the data that populates the list item.
- * @param modifier Modifiers to set for this composable.
- */
-@Composable
-fun BookItem(
-    book: Book,
-    modifier: Modifier = Modifier
-) {
-    // State variable for expanding/collapsing book description
-    var isExpanded by remember { mutableStateOf(false) }
-
-    Card(
-        modifier = modifier
-            .fillMaxWidth()
-            .clickable { isExpanded = !isExpanded } // Toggle isExpanded on click
-            .padding(dimensionResource(R.dimen.padding_small))
-            .animateContentSize() // Animates the content size change
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                // Displays the book icon
-                BookIcon(book.imageResourceId)
-                // Displays book information
-                BookInformation(book.name, book.releaseDate)
-            }
-
-            // Displays the book.description when expanded
-            if (isExpanded) {
-                Text(
-                    text = book.description,
-                    modifier = Modifier.padding(dimensionResource(R.dimen.padding_small))
-                )
-            }
-        }
-    }
-}
 
 
-/**
- * Composable that displays a list item containing a question along with an answer, used for the about us page
- * @param book Contains the data that populates the list item.
- * @param modifier Modifiers to set for this composable.
- */
-@Composable
-fun TextItem(
-    text: About,
-    modifier: Modifier = Modifier
-) {
-    // State variable for expanding/collapsing book description
-    var isExpanded by remember { mutableStateOf(false) }
-
-    Card(
-        modifier = modifier
-            .fillMaxWidth()
-            .clickable { isExpanded = !isExpanded } // Toggle isExpanded on click
-            .padding(dimensionResource(R.dimen.padding_small))
-            .animateContentSize() // Animates the content size change
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                // Displays the book icon
-                Text(text = text.question, fontWeight = FontWeight.Bold, fontSize = 24.sp, modifier = Modifier.padding(dimensionResource(R.dimen.padding_small)))
-            }
-
-            // Displays the book.description when expanded
-            if (isExpanded) {
-                Spacer(modifier = Modifier)
-                Text(
-                    text = text.answer,
-                    modifier = Modifier.padding(dimensionResource(R.dimen.padding_small))
-                )
-            }
-        }
-    }
-}
 
 
-/**
- * Composable function for displaying book information.
- * @param bookName The name of the book.
- * @param releaseDate The release date of the book.
- * @param modifier Modifier for the composable.
- * I commented
- */
-@Composable
-fun BookInformation(
-    bookName: String,
-    releaseDate: Int,
-    modifier: Modifier = Modifier
-) {
-    Column(modifier = modifier) {
-        Text(
-            text = bookName,
-            style = MaterialTheme.typography.titleLarge,
-            modifier = Modifier.padding(top = dimensionResource(R.dimen.padding_small))
-        )
-        Text(
-            text = "Released in: $releaseDate",
-            style = MaterialTheme.typography.bodyLarge
-        )
-    }
-}
+
+
+
+
 
 /**
  * Preview function for MyApp composable.
@@ -499,29 +208,7 @@ fun MyAppPreview() {
     }
 }
 
-/**
- * Composable function for displaying a book icon.
- * @param bookIcon The resource ID of the book icon.
- * @param modifier Modifier for the composable.
- */
-@Composable
-fun BookIcon(
-    @DrawableRes bookIcon: Int,
-    modifier: Modifier = Modifier
-) {
-    Image(
-        modifier = modifier
-            .size(
-                width = dimensionResource(R.dimen.width),
-                height = dimensionResource(R.dimen.height)
-            ) // Sets the size of the image
-            .padding(dimensionResource(R.dimen.padding_small))
-            .clip(MaterialTheme.shapes.small),
-        contentScale = ContentScale.Crop,
-        painter = painterResource(bookIcon),
-        contentDescription = null
-    )
-}
+
 
 /**
  * Composable function for the top app bar.
@@ -571,23 +258,3 @@ fun BookTopAppBar(
     }
 }
 
-@Composable
-fun AboutUs(modifier: Modifier = Modifier) {
-
-    val (textList) = remember { mutableStateOf(questions) }
-
-    Surface(
-        modifier = modifier.fillMaxSize(),
-    ) {
-        LazyColumn(modifier = modifier.fillMaxSize()) {
-
-            items(textList) { question ->
-                // Displays individual book items
-                TextItem(
-                    text = question,
-                    modifier = Modifier.padding(10.dp)
-                )
-            }
-        }
-    }
-}
